deleted uneeded class, database repo refactor, package rename
