+++ cassandra/src/main/java/com/strategicgains/docussandra/event/IndexCreatedEvent.java
@@ -17,7 +17,7 @@ package com.strategicgains.docussandra.event;
 
 import com.strategicgains.docussandra.domain.Identifier;
 import com.strategicgains.docussandra.domain.Index;
-import com.strategicgains.docussandra.domain.abstractparent.Identifiable;
+import com.strategicgains.docussandra.domain.parent.Identifiable;
 import java.io.Serializable;
 import java.util.Date;
 import java.util.List;
